little change

+++ sql/schema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE doings(
-    id bigserial primary key ,
-    chat_id integer not null,
-    name text not null ,
-    importance integer not null ,
-    time timestamp not null
+                       id bigserial primary key ,
+                       chat_id bigint not null,
+                       name text not null ,
+                       importance integer not null ,
+                       time timestamp not null
 )
